refactor(report): forward lint rendering to output
