+++ internal/report/json.go
@@ -3,25 +3,19 @@ Copyright © 2026 envaar
 SPDX-License-Identifier: Apache-2.0
 */
 
-// Package report renders lint findings for humans and machines.
+// Package report is a temporary compatibility facade for lint output.
+//
+// New code should use internal/output/lint directly. The forwarding functions
+// remain until the diff output migration removes this package in #71.
 package report
 
 import (
-	"encoding/json"
-
-	"github.com/envaar/vaar/internal/lint"
+	model "github.com/envaar/vaar/internal/lint"
+	lintoutput "github.com/envaar/vaar/internal/output/lint"
 )
-
-type payload struct {
-	Findings []lint.Finding `json:"findings"`
-}
 
 // JSON renders findings as indented machine-readable output with an explicit
 // findings array, even when the run found nothing.
-func JSON(findings []lint.Finding) ([]byte, error) {
-	if findings == nil {
-		findings = []lint.Finding{}
-	}
-
-	return json.MarshalIndent(payload{Findings: findings}, "", "  ")
+func JSON(findings []model.Finding) ([]byte, error) {
+	return lintoutput.JSON(findings)
 }
